+++ matflow/models/task.py
@@ -220,6 +220,7 @@ class TaskSchema(object):
             'file',
             'default',
             'include_all_iterations',  # inputs from all iterations sent to the input map?
+            'ignore_dependency_from',  # exclude tasks from parameter dependency search
         ]
         req_inp_keys = ['name']
         allowed_inp_keys = req_inp_keys + allowed_inp_specifiers
@@ -236,6 +237,7 @@ class TaskSchema(object):
                 'group': 'default',
                 'file': False,
                 'include_all_iterations': False,
+                'ignore_dependency_from': [],
             }
             inp = get_specifier_dict(inp, name_key='name', defaults=inp_defs)